refactor: ABI
