+++ src/app/collective-rewards/shared/hooks/index.ts
@@ -1,6 +1,7 @@
 export * from './useGetABI'
 export * from './useGetABIWithGraph'
 export * from './useGetABIWithStateSync'
-export * from './useGetCyclePayout'
+export * from './useGetABIFromChain'
 export * from './useGetEstimatedBackersRewardsPct'
 export * from './useReporting'
+export * from './useGetBackerABI'
